+++ commands/generate.go
@@ -303,7 +303,14 @@ func (c *GenerateCommand) generate(apiPaths []swagger.ApiPath) int {
 
 	azapiDefinitionByResourceType := make(map[string][]types.AzapiDefinition)
 	for _, azapiDefinition := range azapiDefinitionsAll {
-		azapiDefinitionByResourceType[azapiDefinition.AzureResourceType] = append(azapiDefinitionByResourceType[azapiDefinition.AzureResourceType], azapiDefinition)
+		azureResourceType := azapiDefinition.AzureResourceType
+		// To avoid the case that there are multiple resource types with the same name but different casing
+		for resourceType, _ := range azapiDefinitionByResourceType {
+			if strings.EqualFold(resourceType, azureResourceType) {
+				azureResourceType = resourceType
+			}
+		}
+		azapiDefinitionByResourceType[azureResourceType] = append(azapiDefinitionByResourceType[azureResourceType], azapiDefinition)
 	}
 
 	resourceTypes := make([]string, 0)